requirements_new.txt
# Telugu S2S System Requirements
# Optimized for H200 training and A6000 inference

# Core Deep Learning
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0

# Transformers & NLP
transformers==4.45.0
tokenizers==0.19.1
datasets==2.18.0
accelerate==0.33.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
pydub==0.25.1
scipy==1.12.0
numpy==1.24.3

# Neural Codec
einops==0.7.0
vector-quantize-pytorch==1.14.24

# Optimization
flash-attn==2.5.0  # Requires CUDA 11.6+
deepspeed==0.14.0
apex  # NVIDIA Apex for mixed precision
triton==2.2.0  # For Flash Attention

# Data Collection
yt-dlp==2024.1.0
whisperx @ git+https://github.com/m-bain/whisperX.git
pyannote.audio==3.1.1

# Web Server
fastapi==0.110.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.9
uvloop==0.19.0  # Faster event loop

# Monitoring & Logging
wandb==0.16.3
tensorboard==2.15.1
tqdm==4.66.1
rich==13.7.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
omegaconf==2.3.0
hydra-core==1.3.2

# Development
ipython==8.20.0
jupyter==1.0.0
black==23.12.1
pylint==3.0.3
pytest==7.4.4

# Deployment
onnx==1.15.0
onnxruntime-gpu==1.17.0
tensorrt==8.6.1  # For NVIDIA optimization

# RunPod specific
runpod==1.4.0

# Optional (for voice cloning future)
# speechbrain==0.5.16
# resemblyzer==0.1.3